+++ src/whipped.rs
@@ -1,2 +1,3 @@
 pub mod whip_s3;
+pub mod whip_az;
 pub mod whip_main;
